feat: fix up get_context
