+++ great_expectations/data_context/data_context/data_context.py
@@ -16,6 +16,9 @@ from great_expectations.data_context.data_context.base_data_context import (
 from great_expectations.data_context.data_context.cloud_data_context import (
     CloudDataContext,
 )
+from great_expectations.data_context.data_context.file_data_context import (
+    FileDataContext,
+)
 from great_expectations.data_context.templates import (
     CONFIG_VARIABLES_TEMPLATE,
     PROJECT_TEMPLATE_USAGE_STATISTICS_DISABLED,
@@ -317,7 +320,7 @@ class DataContext(BaseDataContext):
             )
         else:
             context_root_dir = (
-                self.find_context_root_dir()
+                FileDataContext.find_context_root_dir()
                 if context_root_dir is None
                 else context_root_dir
             )
@@ -471,32 +474,12 @@ class DataContext(BaseDataContext):
         self._save_project_config()
 
     @classmethod
-    def find_context_root_dir(cls) -> str:
-        result = None
-        yml_path = None
-        ge_home_environment = os.getenv("GE_HOME")
-        if ge_home_environment:
-            ge_home_environment = os.path.expanduser(ge_home_environment)
-            if os.path.isdir(ge_home_environment) and os.path.isfile(
-                os.path.join(ge_home_environment, "great_expectations.yml")
-            ):
-                result = ge_home_environment
-        else:
-            yml_path = cls.find_context_yml_file()
-            if yml_path:
-                result = os.path.dirname(yml_path)
-
-        if result is None:
-            raise ge_exceptions.ConfigNotFoundError()
-
-        logger.debug(f"Using project config: {yml_path}")
-        return result
-
-    @classmethod
     def get_ge_config_version(
         cls, context_root_dir: Optional[str] = None
     ) -> Optional[float]:
-        yml_path = cls.find_context_yml_file(search_start_dir=context_root_dir)
+        yml_path = FileDataContext.find_context_yml_file(
+            search_start_dir=context_root_dir
+        )
         if yml_path is None:
             return None
 
@@ -532,7 +515,9 @@ class DataContext(BaseDataContext):
                     ),
                 )
 
-        yml_path = cls.find_context_yml_file(search_start_dir=context_root_dir)
+        yml_path = FileDataContext.find_context_yml_file(
+            search_start_dir=context_root_dir
+        )
         if yml_path is None:
             return False
 
@@ -544,33 +529,6 @@ class DataContext(BaseDataContext):
             yaml.dump(config_commented_map_from_yaml, f)
 
         return True
-
-    @classmethod
-    def find_context_yml_file(
-        cls, search_start_dir: Optional[str] = None
-    ) -> Optional[str]:
-        """Search for the yml file starting here and moving upward."""
-        yml_path = None
-        if search_start_dir is None:
-            search_start_dir = os.getcwd()
-
-        for i in range(4):
-            logger.debug(
-                f"Searching for config file {search_start_dir} ({i} layer deep)"
-            )
-
-            potential_ge_dir = os.path.join(search_start_dir, cls.GE_DIR)
-
-            if os.path.isdir(potential_ge_dir):
-                potential_yml = os.path.join(potential_ge_dir, cls.GE_YML)
-                if os.path.isfile(potential_yml):
-                    yml_path = potential_yml
-                    logger.debug(f"Found config file at {str(yml_path)}")
-                    break
-            # move up one directory
-            search_start_dir = os.path.dirname(search_start_dir)
-
-        return yml_path
 
     @classmethod
     def does_config_exist_on_disk(cls, context_root_dir: str) -> bool:
